Move all merge method for layer into their classes.
